refactor(types) followDomain-Based Type Organization across the project
