+++ src/components/profile/ProfileInfo.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import { mockedAvatar } from '@/assets';
-import { UserProfile } from '@/types/user';
+import { UserProfile } from '@/types';
 import ProfileFieldComponent from './ProfileField';
 import Button from '@/shared/Button/Button';
 import { PROFILE_FIELDS } from '@/constants/profile';
