+++ src/types/models.ts
@@ -1,0 +1,5 @@
+export interface Trainer {
+  id: number;
+  name: string;
+  specialization: string;
+}
