+++ src/types/user.ts
@@ -1,18 +1,18 @@
 export interface UserProfile {
-	firstName?: string;
-	lastName?: string;
-	username?: string;
-	displayName?: string;
-	specialization?: string;
-	address?: string;
-	email?: string;
-	uid?: string;
-	photoURL?: string;
-	status?: string;
+  firstName?: string;
+  lastName?: string;
+  username?: string;
+  displayName?: string;
+  specialization?: string;
+  address?: string;
+  email?: string;
+  uid?: string;
+  photoURL?: string;
+  status?: string;
 }
 
 export interface ProfileField {
-	label: string;
-	key: keyof UserProfile;
-	hidden?: boolean;
+  label: string;
+  key: keyof UserProfile;
+  hidden?: boolean;
 }
